+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added Erlang distribution over serial (uart)
+- Added support for `nif_start`, `executable_line` and `debug_line` opcodes
+- Added named variable debugging support in DWARF when modules are compiled with `beam_debug_info`
 
 ### Fixed
 - Stop using deprecated `term_from_int32` on STM32 platform